Splitted postprocess into the different pipeline stages + minor changes

+++ run_net.lua
@@ -24,7 +24,7 @@ in_im = input.img
 in_im = in_im:view(1,in_im:size(1),in_im:size(2),in_im:size(3))
 in_im = in_im:permute(1,2,4,3)
 in_im:div(255.0):mul(2):add(-1) --Input Normalization
-output = model:forward(in_im:double())
+output = model:forward(in_im:float())
 output:add(1):div(2):mul(255) --Output Denormalization
 output = output:permute(1,2,4,3)
 mattorch.save(params.res, output:double())
